requirements.txt
# DEPENDENCIAS

fastapi
gunicorn
requests
bson
stringcase
uvicorn
pyyaml
dotmap
sqlalchemy